Avoid value truncation warning
STAKING_PORTIONS doesn't fit in a double, and so clang is warning about
the precision loss.  Be explicit about double cast to avoid this
warning.

+++ src/cryptonote_core/service_node_rules.cpp
@@ -183,7 +183,7 @@ static bool get_portions_from_percent(double cur_percent, uint64_t& portions) {
   }
   else
   {
-    portions = (cur_percent / 100.0) * STAKING_PORTIONS;
+    portions = (cur_percent / 100.0) * (double)STAKING_PORTIONS;
   }
 
   return true;
